Exports objects out of files and assemble in crafty.js
This helps to avoid globals and makes it more modular
as the components can be loaded individual.
