+++ src/core/loader.js
@@ -1,7 +1,6 @@
 var Crafty = require('../core/core.js');
 
-
-Crafty.extend({
+module.exports = {
     /**@
      * #Crafty.assets
      * @category Assets
@@ -434,4 +433,4 @@ Crafty.extend({
             }
         }
     }
-});
+};
